+++ react/components/ReviewBlock.tsx
@@ -449,6 +449,7 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
           unitMultiplier: itm?.unitMultiplier,
           moq: itm?.moq,
           refid: itm?.refid,
+          brand: itm?.brand
         }
       })